css: Enhanced vertical rhythm.
Stick harder to a common vertical rhythm. Use LESS to do some of the
hardwork. There are still some elements which are difficult to keep in
the rhythm:

 1. Code listings (this looks odd to have so much space).
 2. <hr> (too difficult to give a size in em since it is a sprite).
 3. Images (they should be padded depending on their sizes but we
    cannot know for sure the size in em unless we use Javascript).
 4. Gallery.

Also, uses a Le Corbusier suite to define headings size
(http://lamb.cc/typograph/). Sizes in em are rounded such that we
achieve pixel perfection when the user has 16px by default. Otherwise,
the vertical rhythm may derive a bit.

Drop h5 and h6 headers (should not be present). Some horizontal
borders have been dropped since the size of the element includes the
size of the borders and therefore, borders should have been included
in vertical rhythm calculations (but this is too difficult).