typography: rework widow plugin
First, when a sentence do not end by a punctuation mark, don't try to
do widow detection. We don't want that for titles or incomplete
sentences (they are usually short).

Second, in French, a double punctuation mark is preceded by a non
breaking space. We need to add it and add one to account for widow
avoidance of the previous space.