nginx: simplify CSP policy
Unfortunately, the use of MathJax in some pages make it quite difficult
to correctly enforce CSP policy. MathJax can use fonts and images to
render the glyphs. We need to search where they are served. Moreover, we
need to accept inline styles. This makes the whole restriction on styles
useless. Instead, just be strict with scripts.