video: don't autoload Youtube JS
Despite using youtube-nocookie.com, Youtube will call DoubleClick which
comes with a cookie for tracking purposes (while the video has no
ads). This is a bit bothersome for users privacy.

Instead, use the cover image (from a cookie-less domain) and if JS is
enabled, replace the image with the actual Youtube video. If no JS, the
cover image is just a link to Youtube (dunno if it would work without
JS).

Some code is inspired by (notably CSS):
 https://webdesign.tutsplus.com/tutorials/how-to-lazy-load-embedded-youtube-videos--cms-26743