images: replace some PNG by SVG
There are fonts. We only use "sans-serif" and "monospace" and try to
correctly set SVG properties to make it work with whatever font we
have. We don't want to force sans-serif since the goal is to avoid load
additional resources and we don't have the appropriate web font for
sans-serif.