images: ensure layout stability when adding captions
Images can be followed by a heading or by text. When followed by a
heading, the heading top margin ensures a minimal distance we have to
match (2 line heights). When we are followed by a text, we don't have
this minimal distance from the text. Therefore, we must ensure that our
outer margin is enough. Without a caption, the outer margin is from
img. With a caption, this is the .lf-captioned paragraph (plus the
caption).