css: increase footnote ref touch surface
On touch device, this is too small to click on it. Use some hack to
expand the surface. We don't overflow vertically to avoid collision with
other links. And we don't overflow too much horizontally (people are
expected to hit the middle, not the border of a link and we shouldn't
have too short links just before).