css: use "system-ui" which should be high quality sans-serif font
For some reason, the default font from browser is usually ancient while
the system-ui font is of top quality.

"system-ui" should work on recent versions of Chrome. Fallback to Roboto
on Android, Segoe UI on Windows, Ubuntu or Cantarell on Linux and
whatever sans-serif font is configured as default font otherwise. Droid
Sans is removed since it is not available since Android 4. Lato is not
widely available as a webfont and is therefore also removed.