extensions: generalize a bit the removal of paragraph
We can do that for a video, a link to a video or an image. We don't do
that for an image since the paragraph is also a signal for us to do a
responsive image which cannot be done at the markdown level since we
need the size of the image (and we also need that for some non-markdown
image).