css: underline links in body text
This is not super pretty on some browsers, but this is better from an
accessibility point of view. I think this is better than border-bottom
as otherwise, the link stands out too much in the text. Outside of the
body text, it is pretty obvious what is a link and what is not.

See: https://www.powermapper.com/products/sortsite/rules/accwcag2-f73-1/