css: switch to Iosevka Term
A custom version with only a subset of characters and with weight
between light and regular. The font still needs to be built. Waiting
on:
 https://github.com/be5invis/Iosevka/issues/238

Inconsolata is great, but has no box drawing characters. Source Code
Pro would be nice but its width is far greater. Iosevka has the same
width ratio than Inconsolata (50%).

The line-height is adapted to match the one from the font. We don't
want overlapping of box elements (but 18px * 1.25 is 22.5px, so
rendering is a bit off on normal DPI screens).