extensions: ensure properties are properties
Some care is needed for an attribute to be a property (without a
value). Parser needs to be set to HTML mode, `set` should be used
directly as the None value is interpreted differently by pyquery. And
there is a trap with `replace_with` that turns its argument as a
string.