build: ensure coherent timestamps between copies
In the past, we didn't copy timestamps because they were always
fresh (we "rm -rf *", then rebuild). However, this means the final
timestamp may differ between servers. Instead, we now ensure
timestamps are accurate with git history, then copy it. The upside is
that copy is faster since we don't need to checksum unmodified files.