css: use font-display swap
Personally, I prefer FOIT to FOUT (I don't like the font change while
I am reading) and I consider the fonts are small, so they should be
served quickly. Moreover, I consider that if FOUT is superior, it
should be the browser default. However, many people advocate for FOUT,
so I am switching to that.