fonts: update Iosevka to latest version
Keep regular x-height. Don't scale more. This is not a perfect match
with Merriweather x-height. We could modify Iosevka to use 575 for
`xheight` and 800 for `cap` to have a better match, but I suppose the
font may not be tuned correctly just for that. The difference is quite
supportable as is.