css: use fallback as a font-display strategy
I wasn't using this because I thought it would be better to not
display anything while the webfont is loading, but in fact, we still
display titles and non-text stuff while the webfont is loading and a
layout shift will occur when webfont is loaded because the metrics are
the ones from the fallback fonts.

Use fallback instead of swap because there is no point in swapping
font if they take too long. Maybe, we can even use optional.