css: do not derive link colors from theme color
Links are meant to be blue and purple, to match the default values and
help users understand the semantic behind them. Therefore, do not
deduce them from the theme color. Also, do not use complex logic to
deduce them one from another.

The dark mode still needs to be tuned a bit, I think the colors are a
bit off.