css: try harder to find a fallback for Iosevka
This is a narrow font. PragmataPro is a good match but not a standard
font. Noto Sans Mono Condensed is a good second and should be
available on Android. Consolas is not bad either and should be
available on Windows. All other fonts are too large.

See https://int08h.com/post/monospace-overchoice/