images: use an iframe to display PDF
This works better for Chromium. It still works in Chrome and Firefox.
If pdf.js is disabled in Firefox, it will propose a download, which is
a pain. However, with the previous method, the fallback link was not
here at all.

We could remove the download prompt using sandboxing, but I didn't
find a way to make it work in Chrome...